+++ Tests/AccessibilityTests/AccessibilityProject/Sources/MainTarget/main.swift
@@ -68,3 +68,7 @@ Task {
 // Property wrappers
 
 _ = PublicWrappedProperty().wrappedProperty
+
+// Inline
+
+inlinableFunction()